+++ fuzz/fuzz_targets/fuzz_target_1.rs
@@ -3,7 +3,7 @@
 use libfuzzer_sys::fuzz_target;
 use rustyfim::NEclatClosed;
 
-fuzz_target!(|transactions: Vec<Vec<usize>>| {
+fuzz_target!(|transactions: Vec<Vec<u32>>| {
     // fuzzed code goes here
     let neclat = NEclatClosed::default();
     neclat.process(&transactions, 0.5);
